Survivors can interact with objects of interest.
The base functionality of allowing the survivors to interact with an object are implemented.

Survivors will be able to interact with an object if they are within the interaction trigger zone AND looking at the interaction object.

Survivors will see a prompt UI informing them to press a key to interact with the type of object.

The next part of this feature will be to actually implement an object in which the survivor can use.